+++ project/src/components/card/card.tsx
@@ -5,8 +5,6 @@ import {useState, useEffect} from 'react';
 import {getCurrentPoint} from '../../store/action';
 import {useAppDispatch} from '../../hooks';
 import {generatePath } from 'react-router';
-import {store} from '../../store';
-import {fetchCommentsAction} from '../../store/api-actions';
 
 type CardProps = {
   card: OfferType;
@@ -49,10 +47,6 @@ function CardScreen(props:CardProps): JSX.Element {
         break;
     }}, [pageType]);
 
-  const getCommentForOffer = () => {
-    store.dispatch(fetchCommentsAction(card.id.toString()));
-    dispatch(getCurrentPoint(card, false));
-  };
 
   return (
     <article
@@ -93,7 +87,7 @@ function CardScreen(props:CardProps): JSX.Element {
           </div>
         </div>
         <h2 className="place-card__name">
-          <Link to={generatePath(`${AppRoute.Property}/:id`, { id: card.id.toString()})} onClick={getCommentForOffer}>{card.title}</Link>
+          <Link to={generatePath(`${AppRoute.Property}/:id`, { id: card.id.toString()})} onClick={() => dispatch(getCurrentPoint(card, false))}>{card.title}</Link>
         </h2>
         <p className="place-card__type">{card.typeOffer}</p>
       </div>
