+++ project/src/components/comment-form/comment-form.tsx
@@ -1,29 +1,32 @@
-import {useRef, FormEvent} from 'react';
+import {useState, FormEvent} from 'react';
 import {useAppDispatch} from '../../hooks';
 import {sendNewComment} from '../../store/api-actions';
-import {store} from '../../store';
-import {fetchCommentsAction} from '../../store/api-actions';
+
 
 type CommentProps = {
   hotelId: number;
 }
 
 function CommentForm({hotelId}:CommentProps): JSX.Element {
+  const EMPTY_COMMENT = {comment: '', rating: 0};
+  const [commentData, setCommentData] = useState(EMPTY_COMMENT);
 
-  const commentRef = useRef<HTMLTextAreaElement | null>(null);
-  const ratingRef = useRef<HTMLInputElement | null>(null);
+  const commentChangeHandle = (evt: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+    const {name, value} = evt.target;
+    setCommentData({...commentData, [name]: value});
+  };
 
   const dispatch = useAppDispatch();
 
   const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
     evt.preventDefault();
-    if (commentRef.current !== null && ratingRef.current !== null) {
+    if (commentData.comment !== '' && commentData.rating !== 0) {
       dispatch(sendNewComment({
-        comment: commentRef.current.value,
-        rating: ratingRef.current.value,
+        comment: commentData.comment,
+        rating: commentData.rating.toString(),
         hotelId: hotelId,
       }));
-      store.dispatch(fetchCommentsAction(hotelId.toString()));
+      setCommentData(EMPTY_COMMENT);
     }
   };
 
@@ -31,35 +34,35 @@ function CommentForm({hotelId}:CommentProps): JSX.Element {
     <form className="reviews__form form" action="" method="post" onSubmit={handleSubmit}>
       <label className="reviews__label form__label" htmlFor="review">Your review</label>
       <div className="reviews__rating-form form__rating">
-        <input className="form__rating-input visually-hidden" name="rating" value="5" id="5-stars" type="radio" ref={ratingRef} />
+        <input className="form__rating-input visually-hidden" name="rating" value="5" id="5-stars" type="radio" checked={commentData.rating === 5} onChange={commentChangeHandle} />
         <label htmlFor="5-stars" className="reviews__rating-label form__rating-label" title="perfect">
           <svg className="form__star-image" width="37" height="33">
             <use xlinkHref="#icon-star"></use>
           </svg>
         </label>
 
-        <input className="form__rating-input visually-hidden" name="rating" value="4" id="4-stars" type="radio" ref={ratingRef} />
+        <input className="form__rating-input visually-hidden" name="rating" value="4" id="4-stars" type="radio" checked={commentData.rating === 4} onChange={commentChangeHandle} />
         <label htmlFor="4-stars" className="reviews__rating-label form__rating-label" title="good">
           <svg className="form__star-image" width="37" height="33">
             <use xlinkHref="#icon-star"></use>
           </svg>
         </label>
 
-        <input className="form__rating-input visually-hidden" name="rating" value="3" id="3-stars" type="radio" ref={ratingRef} />
+        <input className="form__rating-input visually-hidden" name="rating" value="3" id="3-stars" type="radio" checked={commentData.rating === 3} onChange={commentChangeHandle} />
         <label htmlFor="3-stars" className="reviews__rating-label form__rating-label" title="not bad">
           <svg className="form__star-image" width="37" height="33">
             <use xlinkHref="#icon-star"></use>
           </svg>
         </label>
 
-        <input className="form__rating-input visually-hidden" name="rating" value="2" id="2-stars" type="radio" ref={ratingRef} />
+        <input className="form__rating-input visually-hidden" name="rating" value="2" id="2-stars" type="radio" checked={commentData.rating === 2} onChange={commentChangeHandle} />
         <label htmlFor="2-stars" className="reviews__rating-label form__rating-label" title="badly">
           <svg className="form__star-image" width="37" height="33">
             <use xlinkHref="#icon-star"></use>
           </svg>
         </label>
 
-        <input className="form__rating-input visually-hidden" name="rating" value="1" id="1-star" type="radio" ref={ratingRef} />
+        <input className="form__rating-input visually-hidden" name="rating" value="1" id="1-star" type="radio" checked={commentData.rating === 1} onChange={commentChangeHandle} />
         <label htmlFor="1-star" className="reviews__rating-label form__rating-label" title="terribly">
           <svg className="form__star-image" width="37" height="33">
             <use xlinkHref="#icon-star"></use>
@@ -70,7 +73,8 @@ function CommentForm({hotelId}:CommentProps): JSX.Element {
         className="reviews__textarea form__textarea"
         id="comment"
         name="comment"
-        ref={commentRef}
+        value={commentData.comment}
+        onChange={commentChangeHandle}
         placeholder="Tell how was your stay, what you like and what can be improved"
       >
       </textarea>
